// src/lib/apollo/gen/apollostorage/grpc/object_api.ts
export interface PutObjectResponse {
  generation: string;
  crc32c: string;
  md5: string;
  size: string;
}

export interface GetObjectRequest {
  bucket: string;
  object: string;
}

export interface GetObjectResponse {
  header?: ObjectMetadata | undefined;
  chunk?: Buffer | undefined;
}

export interface HeadObjectRequest {
  bucket: string;
  object: string;
}

export interface ObjectMetadata {
  bucket: string;
  object: string;
  contentType: string;
  size: string;
  crc32c: string;
  md5: string;
  generation: string;
}

export interface DeleteObjectRequest {
  bucket: string;
  object: string;
}

export interface DeleteObjectResponse {
  bucket: string;
  object: string;
}
// ...
export const PutObjectResponse: MessageFns<PutObjectResponse> = {
  encode(message: PutObjectResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.generation !== "0") {
      writer.uint32(8).int64(message.generation);
    }
    if (message.crc32c !== "") {
      writer.uint32(18).string(message.crc32c);
    }
    if (message.md5 !== "") {
      writer.uint32(26).string(message.md5);
    }
    if (message.size !== "0") {
      writer.uint32(32).int64(message.size);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PutObjectResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePutObjectResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.generation = reader.int64().toString();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.crc32c = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.md5 = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.size = reader.int64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PutObjectResponse {
    return {
      generation: isSet(object.generation) ? globalThis.String(object.generation) : "0",
      crc32c: isSet(object.crc32c) ? globalThis.String(object.crc32c) : "",
      md5: isSet(object.md5) ? globalThis.String(object.md5) : "",
      size: isSet(object.size) ? globalThis.String(object.size) : "0",
    };
  },

  toJSON(message: PutObjectResponse): unknown {
    const obj: any = {};
    if (message.generation !== "0") {
      obj.generation = message.generation;
    }
    if (message.crc32c !== "") {
      obj.crc32c = message.crc32c;
    }
    if (message.md5 !== "") {
      obj.md5 = message.md5;
    }
    if (message.size !== "0") {
      obj.size = message.size;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<PutObjectResponse>, I>>(base?: I): PutObjectResponse {
    return PutObjectResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<PutObjectResponse>, I>>(object: I): PutObjectResponse {
    const message = createBasePutObjectResponse();
    message.generation = object.generation ?? "0";
    message.crc32c = object.crc32c ?? "";
    message.md5 = object.md5 ?? "";
    message.size = object.size ?? "0";
    return message;
  },
};
// ...
export const ObjectMetadata: MessageFns<ObjectMetadata> = {
  encode(message: ObjectMetadata, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bucket !== "") {
      writer.uint32(10).string(message.bucket);
    }
    if (message.object !== "") {
      writer.uint32(18).string(message.object);
    }
    if (message.contentType !== "") {
      writer.uint32(26).string(message.contentType);
    }
    if (message.size !== "0") {
      writer.uint32(32).int64(message.size);
    }
    if (message.crc32c !== "") {
      writer.uint32(42).string(message.crc32c);
    }
    if (message.md5 !== "") {
      writer.uint32(50).string(message.md5);
    }
    if (message.generation !== "0") {
      writer.uint32(56).int64(message.generation);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): ObjectMetadata {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseObjectMetadata();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.bucket = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.object = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.contentType = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.size = reader.int64().toString();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.crc32c = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.md5 = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.generation = reader.int64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): ObjectMetadata {
    return {
      bucket: isSet(object.bucket) ? globalThis.String(object.bucket) : "",
      object: isSet(object.object) ? globalThis.String(object.object) : "",
      contentType: isSet(object.contentType)
        ? globalThis.String(object.contentType)
        : isSet(object.content_type)
        ? globalThis.String(object.content_type)
        : "",
      size: isSet(object.size) ? globalThis.String(object.size) : "0",
      crc32c: isSet(object.crc32c) ? globalThis.String(object.crc32c) : "",
      md5: isSet(object.md5) ? globalThis.String(object.md5) : "",
      generation: isSet(object.generation) ? globalThis.String(object.generation) : "0",
    };
  },

  toJSON(message: ObjectMetadata): unknown {
    const obj: any = {};
    if (message.bucket !== "") {
      obj.bucket = message.bucket;
    }
    if (message.object !== "") {
      obj.object = message.object;
    }
    if (message.contentType !== "") {
      obj.contentType = message.contentType;
    }
    if (message.size !== "0") {
      obj.size = message.size;
    }
    if (message.crc32c !== "") {
      obj.crc32c = message.crc32c;
    }
    if (message.md5 !== "") {
      obj.md5 = message.md5;
    }
    if (message.generation !== "0") {
      obj.generation = message.generation;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<ObjectMetadata>, I>>(base?: I): ObjectMetadata {
    return ObjectMetadata.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<ObjectMetadata>, I>>(object: I): ObjectMetadata {
    const message = createBaseObjectMetadata();
    message.bucket = object.bucket ?? "";
    message.object = object.object ?? "";
    message.contentType = object.contentType ?? "";
    message.size = object.size ?? "0";
    message.crc32c = object.crc32c ?? "";
    message.md5 = object.md5 ?? "";
    message.generation = object.generation ?? "0";
    return message;
  },
};
// ...
export const DeleteObjectResponse: MessageFns<DeleteObjectResponse> = {
  encode(message: DeleteObjectResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.bucket !== "") {
      writer.uint32(10).string(message.bucket);
    }
    if (message.object !== "") {
      writer.uint32(18).string(message.object);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DeleteObjectResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDeleteObjectResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.bucket = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.object = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DeleteObjectResponse {
    return {
      bucket: isSet(object.bucket) ? globalThis.String(object.bucket) : "",
      object: isSet(object.object) ? globalThis.String(object.object) : "",
    };
  },

  toJSON(message: DeleteObjectResponse): unknown {
    const obj: any = {};
    if (message.bucket !== "") {
      obj.bucket = message.bucket;
    }
    if (message.object !== "") {
      obj.object = message.object;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DeleteObjectResponse>, I>>(base?: I): DeleteObjectResponse {
    return DeleteObjectResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DeleteObjectResponse>, I>>(object: I): DeleteObjectResponse {
    const message = createBaseDeleteObjectResponse();
    message.bucket = object.bucket ?? "";
    message.object = object.object ?? "";
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}